backend/requirements.txt
# Web framework
Django==4.2.16
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1

# CORS
django-cors-headers==4.4.0

# Rate limiting
django-ratelimit==4.1.0

# Database
mysqlclient==2.2.4

# Task queue
celery==5.4.0
redis==5.0.8

# Password hashing
bcrypt==4.2.0
cryptography==42.0.8

# Input sanitisation
bleach==6.1.0

# ML / XAI
shap==0.45.1
scikit-learn==1.5.2
pandas==2.2.3
joblib==1.4.2
numpy==1.26.4

# PDF export
reportlab==4.2.2

# Utilities
python-dotenv==1.0.1
gunicorn==22.0.0
